% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_joint_marginals.R
\name{plot_joint_marginals}
\alias{plot_joint_marginals}
\title{Plot Marginal Densities of g(a) and g(b)}
\arguments{
\item{a_resolution}{Integer. Number of points to use for the uniform density g(a) over [0, 1].}

\item{b_max}{Numeric. Maximum value of b to plot for the gamma-like density g(b).}

\item{b_resolution}{Integer. Number of points to sample along b-axis for plotting g(b).}
}
\value{
A named list containing two `ggplot` objects:
  \describe{
    \item{ga_plot}{Plot of g(a) on [0, 1]}
    \item{gb_plot}{Plot of g(b) on [0, bmax]}
  }
}
\description{
Generates separate `ggplot2` visualizations for the marginal densities
of g(a) and g(b), where a ~ Uniform(0, 1) and b ~ Gamma(2, 1).
}
\examples{
plots <- plot_joint_marginals()
print(plots$ga_plot)
print(plots$gb_plot)
}
