% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmyFApp.R
\name{rmyFApp}
\alias{rmyFApp}
\title{Launch Shiny app for simulating B}
\usage{
rmyFApp()
}
\value{
Launches interactive Shiny app
}
\description{
This function launches a Shiny app that visualizes and summarizes simulations
of B = X + Y using the inverse CDF optimization method.
}
\examples{
rmyFApp()
}
